refactor(web): AmbientBG と Mascot を ~/components/ に共通化
LP からも参照するため chat 配下から共通ディレクトリに移動。
