+++ web/src/app/chat/components/ChatStandingMascot.tsx
@@ -1,7 +1,11 @@
 import { useThreadRuntime } from "@assistant-ui/react";
 import { useEffect, useRef, useState } from "react";
 import { cn } from "~/lib/class-merge";
-import { Mascot, type MascotExpression, type MascotState } from "./Mascot";
+import {
+  Mascot,
+  type MascotExpression,
+  type MascotState,
+} from "~/components/Mascot";
 
 // 応答完了直後にランダムで選ぶ表情。talking 多めで、たまに笑顔・驚きが混じる
 const POST_RESPONSE_STATES: readonly MascotState[] = [
