+++ web/src/app/chat/components/Landing.tsx
@@ -1,7 +1,7 @@
 import { SendIcon } from "lucide-react";
 import { type FormEvent, useState } from "react";
 import { cn } from "~/lib/class-merge";
-import { Mascot } from "./Mascot";
+import { Mascot } from "~/components/Mascot";
 
 const QUICK_PROMPTS: ReadonlyArray<{ icon: string; text: string }> = [
   { icon: "🧭", text: "音威子府村って、どんなところ？" },
