+++ web/src/components/AmbientBG.tsx
