+++ web/src/components/Mascot.tsx
